32_poly:  Split hardware independent function into their own files, for easy testing
