+++ src/lh2/lh2_decoder.h
@@ -1,0 +1,78 @@
+#ifndef __LH2_DECODER_H_
+#define __LH2_DECODER_H_
+
+/**
+ * @defgroup    bsp_lh2 LightHouse 2 decoder algorithm
+ * @ingroup     bsp
+ * @brief       Algorithm to decode the lighthouse pulses, no hardware dependencies.
+ *
+ * @{
+ * @file
+ * @author Said Alvarado-Marin <said-alexander.alvarado-marin@inria.fr>
+ * @copyright Inria, 2025-present
+ * @}
+ */
+
+#include <stdint.h>
+#include <stdbool.h>
+#include "lh2.h"
+
+//=========================== defines ==========================================
+
+// Dynamic checkpoints for the lsfr index search
+typedef struct {
+    uint32_t bits[LH2_POLYNOMIAL_COUNT][LH2_SWEEP_COUNT];   ///< lfsr pseudo-random bits of the checkpoints
+    uint32_t count[LH2_POLYNOMIAL_COUNT][LH2_SWEEP_COUNT];  ///< corresponding lfsr index of the checkpoints
+} _lfsr_checkpoint_t;
+
+//=========================== public ===========================================
+/**
+ * @brief
+ * @param[in] sample_buffer: SPI samples loaded into a local buffer
+ * @return chipsH: 64-bits of demodulated data
+ */
+uint64_t _demodulate_light(uint8_t *sample_buffer);
+
+/**
+ * @brief from a 17-bit sequence and a polynomial, generate up to 64-17=47 bits as if the LFSR specified by poly were run forwards for numbits cycles, all little endian
+ *
+ * @param[in] poly:     17-bit polynomial
+ * @param[in] bits:     starting seed
+ * @param[in] numbits:  number of bits
+ *
+ * @return sequence of bits resulting from running the LFSR forward
+ */
+uint64_t _poly_check(uint32_t poly, uint32_t bits, uint8_t numbits);
+
+/**
+ * @brief find out which LFSR polynomial the bit sequence is a member of
+ *
+ * @param[in] chipsH1: input sequences of bits from demodulation
+ * @param[in] start_val: number of bits between the envelope falling edge and the beginning of the sequence where valid data has been found
+ *
+ * @return polynomial, indicating which polynomial was found, or FF for error (polynomial not found).
+ */
+uint8_t _determine_polynomial(uint64_t chipsH1, int8_t *start_val);
+
+/**
+ * @brief counts the number of 1s in a 64-bit
+ *
+ * @param[in] bits_in: arbitrary bits
+ *
+ * @return cumulative number of 1s inside of bits_in
+ */
+uint64_t _hamming_weight(uint64_t bits_in);
+
+/**
+ * @brief finds the position of a 17-bit sequence (bits) in the sequence generated by polynomial3 with initial seed 1
+ *
+ * @param[in] checkpoints: structure with dynamic checkpoints
+ * @param[in] index: index of polynomial
+ * @param[in] bits: 17-bit sequence
+ *
+ * @return count: location of the sequence
+ */
+uint32_t _lfsr_index_search(_lfsr_checkpoint_t *checkpoints, uint8_t index, uint32_t bits);
+
+
+#endif /* __LH2_DECODER_H_ */
